requirements.txt
# requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-socketio==5.8.0
aiofiles==23.1.0
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.0
itsdangerous==2.2.0
httpx==0.24.1
python-multipart==0.0.6
pytest-asyncio==0.21.1
# --- THIS IS THE FIX: Use stable versions of passlib and add bcrypt explicitly ---
passlib==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
email-validator==2.1.1